requirements.txt
matplotlib==3.10.8
numpy==2.4.2
pyserial==3.5
scipy==1.17.1
sounddevice==0.5.5
